php Captcha code with demo
We will create an image tag and set the src to core/captcha_image. php.
We will create a form for the user to fill in the text from the image they see.
We will process the form input as soon as the user hits the submit button, then show the result of our verification to the user.